Fiks struktur mellom interfaces og service impl i service og alle interfaces i domain
