--- client/src/main/kotlin/no/nav/su/se/bakover/client/antivirus/VirusScanRequest.kt
@@ -1,7 +1,0 @@
-package no.nav.su.se.bakover.client.antivirus
-
-@Suppress("ArrayInDataClass")
-data class VirusScanRequest(
-    val tittel: String,
-    val fil: ByteArray,
-)
