+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/fritekst/FritekstService.kt
@@ -1,8 +1,6 @@
 package no.nav.su.se.bakover.domain.fritekst
 
 import arrow.core.Either
-import arrow.core.left
-import arrow.core.right
 import no.nav.su.se.bakover.common.persistence.SessionContext
 import java.util.UUID
 
@@ -12,36 +10,6 @@ interface FritekstService {
     fun hentFritekst(referanseId: UUID, type: FritekstType, sessionContext: SessionContext? = null): Either<FritekstFeil, Fritekst>
     fun lagreFritekst(fritekst: FritekstDomain): Unit
     fun slettFritekst(referanseId: UUID, type: FritekstType, sakId: UUID): Either<FritekstFeil, Unit>
-}
-
-interface FritekstRepo {
-    fun hentFritekst(referanseId: UUID, type: FritekstType, sessionContext: SessionContext? = null): Fritekst?
-    fun lagreFritekst(fritekst: Fritekst)
-    fun slettFritekst(referanseId: UUID, type: FritekstType)
-}
-
-class FritekstServiceImpl(
-    private val repository: FritekstRepo,
-) : FritekstService {
-    override fun hentFritekst(hentDomain: FritekstHentDomain): Either<FritekstFeil, Fritekst> {
-        return hentFritekst(
-            referanseId = hentDomain.referanseId,
-            type = hentDomain.type,
-        )
-    }
-
-    override fun hentFritekst(referanseId: UUID, type: FritekstType, sessionContext: SessionContext?): Either<FritekstFeil, Fritekst> {
-        val fritekst = repository.hentFritekst(referanseId = referanseId, type = type, sessionContext = sessionContext)
-        return fritekst?.right() ?: FritekstFeil.FantIkkeFritekst.left()
-    }
-
-    override fun lagreFritekst(fritekst: FritekstDomain) {
-        return repository.lagreFritekst(fritekst.toFritekst())
-    }
-
-    override fun slettFritekst(referanseId: UUID, type: FritekstType, sakId: UUID): Either<FritekstFeil, Unit> {
-        return repository.slettFritekst(referanseId, type).right()
-    }
 }
 
 data class FritekstHentDomain(
