+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/notat/VedleggRepo.kt
@@ -1,0 +1,10 @@
+package no.nav.su.se.bakover.domain.notat
+
+import java.util.UUID
+
+interface VedleggRepo {
+    fun leggTil(vedlegg: NotatVedlegg)
+    fun slett(vedleggId: UUID)
+    fun hent(vedleggId: UUID): NotatVedlegg?
+    fun hentForNotat(notatId: UUID): List<NotatVedlegg>
+}
