+++ service/src/test/kotlin/no/nav/su/se/bakover/service/antivirus/DefaultVirusScanServiceTest.kt
@@ -6,7 +6,8 @@ import no.nav.su.se.bakover.client.antivirus.ClamAVClient
 import no.nav.su.se.bakover.client.antivirus.ScanResponse
 import no.nav.su.se.bakover.client.antivirus.ScanResult
 import no.nav.su.se.bakover.client.antivirus.ScanStatus
-import no.nav.su.se.bakover.client.antivirus.VirusScanRequest
+import no.nav.su.se.bakover.domain.antivirus.VirusScanRequest
+import no.nav.su.se.bakover.domain.antivirus.VirusScanResult
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.mockito.kotlin.any
